requirements.txt
streamlit
openai
pypdf2
tiktoken
pandas
python-docx
openpyxl
langchain



